Core/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
colorama==0.4.6
constant==0.0.4
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==41.0.3
customtkinter==5.2.1
darkdetect==0.8.0
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==6.0.0
Django==4.2.1
django-allauth==0.55.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-environ==0.11.2
django-filter==24.2
django-formtools==2.4.1
django-jazzmin==2.6.0
django-jet-reboot==1.3.3
django-multiselectfield==0.1.12
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.4.1
django-two-factor-auth==1.15.4
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
Faker==19.3.0
feedparser==6.0.10
filelock==3.12.3
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.1.1
numpy==1.26.3
oauthlib==3.2.2
openai==0.27.9
packaging==23.1
phonenumbers==8.13.20
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.21.12
psycopg==3.1.10
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python3-openid==3.2.0
pythonfinder==2.0.5
pytz==2023.3
qrcode==7.4.2
redis==5.0.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
tkcalendar==1.6.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
virtenv==0.3.1
virtualenv==20.24.3
yarl==1.9.2
